Move Santa classes in dedicated package
